Increased Dispatcher efficiency
